refactor(hush-cli): extract search from pkg_cli
Move cmd_pkg_search into pkg_cli/search.rs, widened to pub(super).
search.rs pulls truncate_with_ellipsis/urlencoding_simple directly
from super::util (their pub(super) home) since truncate_with_ellipsis
is now only a test-time import in mod.rs.
